Only process Shaders on Pipeline change
